chore: 이미지 변경 및 프로젝트 진행 상태 삭제
