+++ src/components/common/ProjectsStatusCards.tsx
@@ -5,7 +5,7 @@ import StatusCard from "./ProjectsStatusCard";
 import { ReactNode, useEffect, useState } from "react";
 import axios from "axios";
 import { Loading } from "./Loading";
-import { Folder, Signature, Skull, Swords, Wrench } from "lucide-react";
+import { Folder, PackageCheck, Signature, Swords, Wrench } from "lucide-react";
 
 interface StatusCardsProps {
   title: string;
@@ -33,7 +33,7 @@ const StatusCards: React.FC<StatusCardsProps> = ({ title }) => {
           <Folder key="1" size={40} color="gray" />,
           <Signature key="2" size={40} color="gray" />,
           <Swords key="3" size={40} color="gray" />,
-          <Skull key="4" size={40} color="gray" />,
+          <PackageCheck key="4" size={40} color="gray" />,
           <Wrench key="5" size={40} color="gray" />,
         ];
 
@@ -93,8 +93,7 @@ const StatusCards: React.FC<StatusCardsProps> = ({ title }) => {
         border="1px solid"
         borderColor="gray.200"
         borderRadius="lg"
-        boxShadow="md"
-      >
+        boxShadow="md">
         {data.map((item, index) => (
           <StatusCard
             key={index}
