+++ src/components/pages/ProjectsPage/index.tsx
@@ -1,5 +1,5 @@
-import { Table } from "@chakra-ui/react";
-import BasicTable from "@/src/components/common/BasicTable";
+import { Heading, Table } from "@chakra-ui/react";
+import BasicTable from "@/src/components/common/ProjectsList";
 import ProjectStatusCards from "@/src/components/common/ProjectsStatusCards";
 import Head from "next/head";
 import { ProjectsFilterProvider } from "@/src/context/ProjectsFilterContext";
@@ -17,6 +17,9 @@ export default function Home() {
         />
       </Head>
       <ProjectStatusCards title={"프로젝트 현황"} />
+      <Heading size="2xl" color="gray.600">
+        프로젝트 목록
+      </Heading>
       <BasicTable
         headerTitle={
           <Table.Row
@@ -26,13 +29,11 @@ export default function Home() {
                 // 모든 자식 `th` 태그에 스타일 적용
                 textAlign: "center",
               },
-            }}
-          >
+            }}>
             <Table.ColumnHeader>프로젝트명</Table.ColumnHeader>
             <Table.ColumnHeader>고객사</Table.ColumnHeader>
             <Table.ColumnHeader>개발사</Table.ColumnHeader>
             <Table.ColumnHeader>프로젝트 상태</Table.ColumnHeader>
-            <Table.ColumnHeader>진행 단계</Table.ColumnHeader>
             <Table.ColumnHeader>프로젝트 시작일</Table.ColumnHeader>
             <Table.ColumnHeader>프로젝트 종료일</Table.ColumnHeader>
           </Table.Row>
