+++ src/types/project.ts
@@ -4,7 +4,6 @@ export interface ProjectProps {
   client: string;          // 고객사 이름
   developer: string;       // 개발사 이름
   projectStatus: string;   // 계약 단계
-  progressStepName: string;   // 진행 단계
   startAt: string;       // 시작일
   closeAt: string;         // 종료일
 }
